Move api controllers into an api-specific namespace
