+++ app/Http/Controllers/Api/CreatureTypesController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use App\Models\Data\CreatureType;
 use Illuminate\Http\Request;
