+++ app/Http/Controllers/Api/EditionsController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use App\Models\Data\Edition;
 use Illuminate\Http\Request;
