+++ app/Http/Controllers/Api/ModulesController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use App\Models\Data\Module;
 use App\Models\ModuleListFilter;
 use Illuminate\Http\Request;
