+++ app/Http/Controllers/Api/SettingsController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use App\Models\Data\Setting;
 use Illuminate\Http\Request;
